goji: new SubMux function; be less magical
Previously, Goji would automatically extract a path from the incoming
URL if one was not already present, but would use an existing path
otherwise (for instance, if one was set by a "parent" Mux). This,
combined with cooperation from Pattern implementations, meant that
"sub-router" behavior would work automatically (and magically) out of
the box.

However, several early users found this behavior unintuitive and had
difficulty finding documentation explaining it. Furthermore, the fact
that Mux's behavior was context-dependent and magical was against the
spirit of the project.

Introducing a new SubMux function not only makes what was once implicit
explicit, it provides a useful place to document Goji's sub-router
behavior in greater detail.

This is a breaking API change: users who use Goji sub-routers will have
to replace some calls to NewMux with calls to SubMux in order to
explicitly annotate when routing information should and should not be
inherited.